backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
alembic==1.12.1
sentence-transformers==2.2.2
numpy==1.24.3
torch==2.1.1
openai==1.3.5 